src/backend/requirements.txt
# Core dependencies
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6,<1.0.0

# Database
asyncpg>=0.29.0,<1.0.0
sqlalchemy[asyncio]>=2.0.25,<3.0.0
alembic>=1.13.0,<2.0.0

# Azure
azure-identity>=1.15.0,<2.0.0
azure-keyvault-secrets>=4.7.0,<5.0.0
azure-storage-blob>=12.19.0,<13.0.0
azure-data-tables>=12.5.0,<13.0.0
azure-monitor-opentelemetry>=1.2.0,<2.0.0
azure-communication-sms>=1.0.0,<2.0.0
azure-communication-email>=1.0.0,<2.0.0

# AI / Agent Framework
azure-ai-projects>=1.0.0b1

# HTTP Client
httpx>=0.26.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Utilities
python-dateutil>=2.8.0,<3.0.0
orjson>=3.9.0,<4.0.0

# Observability
opentelemetry-api>=1.22.0,<2.0.0
opentelemetry-sdk>=1.22.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.43b0
structlog>=24.1.0,<25.0.0
